updated instructions for release 2.0.0
Made it more easier to add /remove modules. Previously the class-name was required and the functions were hidden.
Now only the filename is required and the functions are visible.
The section how to set up this module has been removed since inkycal-modules should pass their required and optional config to the web-ui instead of asking the user.
JSON files are always a nice thing, but asking someone without any prior knowledge might be too much to ask.